pyproject.toml
[tool.poetry]
name = "git-llm-tool"
version = "0.1.0"
description = "AI-powered git commit message and changelog generator"
authors = ["skyler-gogolook <skyler.lo@gogolook.com>"]
readme = "README.md"
packages = [{include = "git_llm_tool"}]

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.0"
pyyaml = "^6.0"
openai = "^1.0.0"
anthropic = "^0.20.0"
google-generativeai = "^0.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-mock = "^3.14.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"

[tool.poetry.scripts]
git-llm = "git_llm_tool.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]